chore: remove the old sidebar
