+++ ui/components/ui/index.ts
@@ -10,5 +10,4 @@ export * from "./label/Label";
 export * from "./main-layout/main-layout";
 export * from "./select/Select";
 export * from "./sidebar";
-export * from "./sidebar-new/sidebar";
 export * from "./toast";
