--- ui/components/ui/sidebar-new/index.ts
@@ -1,5 +1,0 @@
-export * from "./collapse-menu-button";
-export * from "./menu";
-export * from "./sheet-menu";
-export * from "./sidebar";
-export * from "./sidebar-toggle";
