+++ ui/components/ui/sidebar/index.ts
@@ -1,6 +1,5 @@
+export * from "./collapse-menu-button";
+export * from "./menu";
+export * from "./sheet-menu";
 export * from "./sidebar";
-export * from "./sidebar-items";
-export * from "./sidebar-wrap";
-export * from "./skeleton-profile";
-export * from "./team-avatar";
-export * from "./user-avatar";
+export * from "./sidebar-toggle";
